% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_num_zonalStats.R
\name{append_num_zonalStats}
\alias{append_num_zonalStats}
\title{Append zonal statistics column(s) to vector data/}
\usage{
append_num_zonalStats(
  input,
  raster,
  funs = c("mean", "min", "max", "quantile", "nodata", "propNA"),
  quantiles = c(0.5, 0.9, 0.95),
  ...
)
}
\arguments{
\item{input}{Vector data of type \code{sf} or \code{sfc} (from \link[sf:sf-package]{sf::sf-package} package) such as polygon or point data.}

\item{raster}{A \link[terra:SpatRaster-class]{terra::SpatRaster} object created by the \link[terra:terra-package]{terra::terra-package} package or \code{raster} object created by the \code{raster} package.
NOTE: \code{raster} is expected to be \href{https://r-spatial.org/r/2022/04/12/evolution.html#packages-depending-on-sp-and-raster}{deprecated in the future}, so using \link[terra:terra-package]{terra::terra-package} is recommended.}

\item{funs}{A vector of zonal statistic functions passed to the \code{fun} argument in \code{\link[exactextractr:exact_extract]{exactextractr::exact_extract()}}. May be either names of known functions described in \code{\link[exactextractr:exact_extract]{exactextractr::exact_extract()}} or (named) custom functions that expect an argument 'df'. See examples for more detail.}

\item{quantiles}{Numeric vector between 0 and 1. When \code{funs} includes the "quantile" zonal statistic function, which quantiles should be extracted.}

\item{...}{Additional arguments passed to \code{\link[exactextractr:exact_extract]{exactextractr::exact_extract()}} function.}
}
\value{
Object of type \code{sf}
}
\description{
Wrapper around \code{exactextractr} function that overcomes some minor issues in the original function.
\enumerate{
\item Takes vector data as first argument to allow function to be more easily used in piped spatial workflow.
\item A non-destructive method that returns a spatial object with all columns by default.
}
}
\examples{
### EXAMPLE 1: USE SINGLE KNOWN ZONAL STAT FUNCTION

## Load a raster file from the terra package
## This is a digital elevation map of Luxembourg
raster <- terra::rast(system.file("ex/elev.tif", package="terra"))

## Polygon objects
## Use data from terra package
polygon_sf <- sf::read_sf(system.file("ex/lux.shp", package="terra"))
multipolygon_sf <- polygon_sf |>
    dplyr::group_by(NAME_1) |>
    dplyr::summarise()

## Find the max value for each of these objects
## NOTE: The zonal statistic is added as a new column to the original dataset
append_num_zonalStats(polygon_sf, raster, funs = "max")
append_num_zonalStats(multipolygon_sf, raster, funs = "max")

### EXAMPLE 2: USE MULTIPLE ZONAL STAT FUNCTIONS

## Find max, min, q50, and q75 elevation within each polygon
## NOTE: Again, all columns are added to the existing dataset
append_num_zonalStats(polygon_sf, raster, funs = c("max", "min", "quantile"),
                quantile = c(0.5, 0.75))
append_num_zonalStats(multipolygon_sf, raster, funs = c("max", "min", "quantile"),
                quantile = c(0.5, 0.75))

### EXAMPLE 3: USE CUSTOM ZONAL STATS FUNCTION

## Find the range of elevation within each polygon
## NOTE: The function must take df and ...
append_num_zonalStats(polygon_sf, raster,
                funs = c("range" = function(df, ...){
max(df$value, na.rm = TRUE) - min(df$value, na.rm = TRUE)}))
}
