% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arg_type.R
\name{check_arg_type}
\alias{check_arg_type}
\title{Test argument type}
\usage{
check_arg_type(
  arg,
  expected_types,
  arg_name,
  msg = "'\%s' must be of type \%s",
  allow_null = FALSE
)
}
\arguments{
\item{arg}{The argument to check. Can be a single value or a named list of values to check.}

\item{expected_types}{A character vector of expected types (e.g., c("numeric", "integer")).
When more than two types are provided, arg may be EITHER type but doesn't have to
be both. If a named list is provided for 'arg' all elements of the list are expected
to fulfill these expected types.}

\item{arg_name}{The name of the argument (for error message). Ignored if a named
list is provided.}

\item{msg}{Character. Message to be printed if type doesn't match. This string
will be passed to \code{sprintf()}. Requires placeholders for two strings that will
be replaced with 'arg_name' (or name from list) and 'expected_types'}

\item{allow_null}{Logical. If arg is not of the expected type, can it still pass
the check if it is NULL. Default (FALSE), arg must have a value that matches
expected_types. If TRUE, 'arg' may be NULL.}
}
\value{
NULL if check passes, otherwise throws an error
}
\description{
A utility function to check if arguments are of the expected type(s) and return meaningful error messages.
}
\examples{
check_arg_type(5, "numeric", "my_arg") ## Works
#check_arg_type(1, "character", "my_arg") ## Fails
#check_arg_type(1, c("character", "factor"), "my_arg", msg = "'\%s' IS WRONG. NOT \%s!!")
#check_arg_type(list(NUM = 1, NUM2 = 6, NUM3 = "X"), c("numeric"), "my_args") ## Fails
## If we just provide an unnamed list we can test for type list.
check_arg_type(list(1, 2, "text"), "list", "my_args")
}
