% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_num_riskValue.R
\name{trans_num_riskValue}
\alias{trans_num_riskValue}
\title{Convert numeric data into 1-5 risk values.}
\usage{
trans_num_riskValue(input, breaks, fn, direction = 1)
}
\arguments{
\item{input}{Numeric vector. Original state of nature data to convert.}

\item{breaks}{Numeric vector length 6. In general, it is good to use -Inf and Inf
as the minimum/maximum values to avoid issues when comparing decimals.}

\item{fn}{List of functions (length 5) to create breaks.
By default, breaks are applied as: <=, <=, <=, <=, >}

\item{direction}{Integer. Do low values in input correspond to
low risk (direction = 1) or high risk (direction = -1)?}
}
\value{
A numeric vector
}
\description{
Intended for use within pipes and can be applied across multiple columns.
See example.
}
\examples{
## Apply breaks to iris data as an example
my_breaks <- c(-Inf, 4.5, 5.5, 6.5, 7.5, Inf)

## Apply to one column using dplyr & pipe
iris |>
  mutate(riskValue = trans_num_riskValue(Sepal.Length, my_breaks))

## Apply to multiple columns using across
iris |>
  mutate(across(.cols = c(Sepal.Length, Sepal.Width),
                \(x) trans_num_riskValue(x, breaks = my_breaks,
                                         direction = -1)))
}
