% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_poly_cutDateline.R
\name{trans_poly_cutDateline}
\alias{trans_poly_cutDateline}
\title{Cut polygons that cross dateline.}
\usage{
trans_poly_cutDateline(sf, crs = "EPSG:4326")
}
\arguments{
\item{sf}{Polygon data of class sf}

\item{crs}{Character string. The CRS of the sf data (currently only EPSG:4326)
is possible.}
}
\value{
Polygon data of class sf
}
\description{
Unproject polygon (EPSG:4326) that cross dateline can be broken when
transforming. This function takes identifies these polygons and splits them
into two to allow transformation.
}
\examples{
fiji <- sf::st_read(system.file("extdata/fiji.gpkg", package = "RFSUtils"))

## Fiji is one polygon but crosses dateline
nrow(fiji)
plot(fiji)

## Split across dateline
fiji_split <- fiji |>
  trans_poly_cutDateline()

## Now we have two polygons either side
nrow(fiji_split)
plot(fiji_split[1, ])
plot(fiji_split[2, ])
}
